Don't output holes that don't have NOTE_ATTACK / OFF_TIME
These keys are either both present, or both absent (in the rolls used to date, at least) -- if they're absent, we don't have a hole to highlight, so there's no need to be loading the data into the app.  Dropping these unusable data makes the files ~60% smaller to drag over the wire (before compression, at least).

I've not investigated thoroughly yet -- are these the "continuation" holes, where a not is "held down", and the duration (NOTE_ATTACK and OFF_TIME) are recorded for the first hole only?